+++ sdk/communication/azure-communication-sms/azure/communication/sms/__init__.py
@@ -3,7 +3,7 @@ from ._delivery_reports_client import DeliveryReportsClient
 from ._opt_outs_client import OptOutsClient
 from ._telco_messaging_client import TelcoMessagingClient
 
-from ._models import SmsSendResult
+from ._models import SmsSendResult, OptOutResult, OptOutCheckResult
 
 __all__ = [
     "SmsClient",
@@ -11,4 +11,6 @@ __all__ = [
     "OptOutsClient",
     "TelcoMessagingClient",
     "SmsSendResult",
+    "OptOutResult",
+    "OptOutCheckResult",
 ]